shutdown: Do 'time moving backward' check only on Reboot
Running FullReboot and Direct_EC_Reboot on Chromebook with no battery
resets time and thus triggers 'time moving backward'. Since this is an
expected result, we skip 'time moving backward' check on FullReboot
and Direct_EC_Reboot and only check when running Reboot.

In addition, there is no battery for Chromebox and the tests should
pass when running on a Chromebox.

BUG=b:203164468

TEST=make test
TEST=run FullReboot on Chromebook and Chromebox
TEST=run Reboot on Chromebook and Chromebox
TEST=run Direct_ec_Reboot on Chromebook and Chromebox

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3347194
Tested-by: Tzu-Min Sun <jimmysun@chromium.org>
Commit-Queue: Tzu-Min Sun <jimmysun@chromium.org>
(cherry picked from commit ec40ea7dbe396d4fd29451edd0de16b5ede80252)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4936480
Commit-Queue: Derek Huang <derekhuang@google.com>
Tested-by: Assion Lee <assion.lee@lcfc.corp-partner.google.com>

+++ py/test/pytests/shutdown.py
@@ -254,8 +254,12 @@ class ShutdownTest(test_case.TestCase):
 
     now = time.time()
     logging.info('%.03f s passed since reboot', now - last_shutdown_time)
-
-    if last_shutdown_time > now:
+    if (self.args.operation == SHUTDOWN_TYPES.reboot and
+        last_shutdown_time > now):
+      # We only check this for reboot. Running fullreboot and direct_ec_reboot
+      # on a Chromebook with no battery resets time and thus triggers the
+      # 'time moving backward' error. Since this is an expected result, we skip
+      # this check for fullreboot and direct_ec_reboot.
       LogAndEndTest(status=state.TestState.FAILED,
                     error_msg=('Time moved backward during reboot '
                                '(before=%s, after=%s)' %
